pyproject.toml
[tool.poetry]
name = "sherpa"
version = "0.1.0"
description = "A CLI tool for loading GIS files to a PostGIS database"
license = "MIT"
authors = ["Connor <contact@kyoh.run>"]
readme = "README.md"
repository = "https://github.com/kyoh-dev/sherpa"

[tool.poetry.scripts]
sherpa = "sherpa.main:app"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "2.9.3"
typer = {version = "^0.4.1", extras = ["all"]}
colorama = "^0.4.5"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
mypy = "^0.961"
autoflake = "^1.4"
black = "^22.6.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
install_types = true
non_interactive = true
ignore_missing_imports = true
show_error_codes = true

[tool.black]
line-length = 100
target-version = ['py39']
